demos/2_mcp_with_crewai/requirements.txt
crewai
crewai-tools

openai==1.75.0
openai-agents==0.0.12

mcp==1.6.0
mcpadapt==0.1.3

torch==2.7.0
accelerate==1.6.0
llama-cpp-python==0.3.8

uv==0.6.14
